+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 1.0.1
 
 - updated platform support specification
+
+## 1.0.2
+
+- added implementation of hashCode getter and == operator
+- added implementation for unwrapOrElse 
+- added implementation for unwrapOrElseAsync